client/source/requirements.txt
pwntools
pycryptodome
requests
flask